// frontend/src/context/authContext/authContext.types.ts
// src/contexts/AuthContext/AuthContext.types.ts

export interface User {
  id: string;
  name: string;
  email: string;
}

export interface AuthContextType {
  user: User | null;
  login: (email: string, password: string) => Promise<void>;
  signup: (email: string, password: string) => Promise<void>;
  logout: () => void;
  isLoading: boolean;
  error: string | null;
  refreshSession: () => Promise<void>;
}